modularized and imported electron's application menu
