--- client/src/application-menu.ts
@@ -1,47 +1,0 @@
-// CURRENTLY NOT USED.
-// NEED TO IMPORT TO ELECTRON.TS
-
-// import { Menu } from 'electron';
-
-export const template = [
-	{
-		label: 'File',
-		submenu: [
-			{
-				label: 'Create File'
-			},
-			{
-				label: 'Create Notebook'
-			},
-			{
-				label: 'Open Notebook'
-			},
-			{
-				label: 'Save'
-			},
-			{
-				label: 'Save As'
-			},
-			{
-				label: 'Export'
-			},
-			{
-				label: 'Exit'
-			}
-		]
-	},
-	{
-		label: 'Edit',
-		submenu: [
-			{
-				label: 'Copy'
-			},
-			{
-				label: 'Pase'
-			},
-			{
-				label: 'Cut'
-			}
-		]
-	}
-];
